Cleanup after merge. Still learning Git. :P
Cleaned up some traces and errors in the PCB editor. Most of the changes won't actually be visible on a manufactured board but cleans things up for anyone who wants to poke around in the editor.

Nudged some solder masks and traces around to remove overlaps. and cleaned up copper zones making many of the runs smoother and look better but should not any real impact on the finished product.

I noticed that the physical board that I received has a "P6-Alt" connector that was not present in the editor so I added that.

Added Redherring32's NullCIC design built into the main board instead of as a daughter board. You can choose to either populate the CIC or the capacitor, resistor, and IC under it.

Updated schematic to reflect changes to the PCB